douyin/requirements_douyin.txt
aiohttp>=3.9.0
aiofiles>=23.2.0
openai-whisper>=20231117
groq>=0.4.0
ffmpeg-python>=0.2.0
